refactor(ChessGameInfo): DB의 chess_game을 객체로 받는 ChessGameInfo 클래스 생성
